@@ -39,7 +39,7 @@ setup(
     long_description_content_type='text/markdown',
     install_requires=['protobuf>=3.11.3,<4.0.0'],
     tests_require=['pytest'],
-    python_requires='>=3.5',
+    python_requires='>=3.8',
     classifiers=[
         'License :: OSI Approved :: Apache Software License',
         'Programming Language :: Python :: 3.5',
