@@ -15,160 +15,134 @@
 #  See the License for the specific language governing permissions and
 # limitations under the License.
 ################################################################################
-
-from google.protobuf.any_pb2 import Any
-import inspect
-
-from enum import Enum
-from typing import List
+import abc
+from dataclasses import dataclass
 from datetime import timedelta
-
-from statefun.kafka_egress_pb2 import KafkaProducerRecord
-from statefun.kinesis_egress_pb2 import KinesisEgressRecord
-
-class SdkAddress(object):
-    def __init__(self, namespace, type, identity):
-        self.namespace = namespace
-        self.type = type
-        self.identity = identity
-
-    def __repr__(self):
-        return "%s/%s/%s" % (self.namespace, self.type, self.identity)
-
-    def typename(self):
-        return "%s/%s" % (self.namespace, self.type)
+from keyword import iskeyword, kwlist
 
 
-class AnyStateHandle(object):
-    def __init__(self, any_bytes):
-        self.any = None
-        self.value_bytes = any_bytes
-        self.modified = False
-        self.deleted = False
-
-    #
-    # TODO This should reflect the actual type URL.
-    # TODO we can support that only after reworking the SDK.
-    #
-    def typename(self):
-        return "type.googleapis.com/google.protobuf.Any"
-
-    def bytes(self):
-        if self.deleted:
-            raise AssertionError("can not obtain the bytes of a delete handle")
-        if self.modified:
-            return self.value.SerializeToString()
-        else:
-            return self.value_bytes
-
-    def unpack(self, into_class):
-        if self.value:
-            into_ref = into_class()
-            self.value.Unpack(into_ref)
-            return into_ref
-        else:
-            return None
-
-    def pack(self, message):
-        any = Any()
-        any.Pack(message)
-        self.value = any
-
-    @property
-    def value(self):
-        """returns the current value of this state"""
-        if self.deleted:
-            return None
-        if self.any:
-            return self.any
-        if not self.value_bytes:
-            return None
-        self.any = Any()
-        self.any.ParseFromString(self.value_bytes)
-        return self.any
-
-    @value.setter
-    def value(self, any):
-        """updates this value to the supplied value, and also marks this state as modified"""
-        self.any = any
-        self.value_bytes = None
-        self.modified = True
-        self.deleted = False
-
-    @value.deleter
-    def value(self):
-        """marks this state as deleted and also as modified"""
-        self.any = None
-        self.value_bytes = None
-        self.deleted = True
-        self.modified = True
+@dataclass(repr=True, eq=True, order=False, frozen=True)
+class SdkAddress:
+    namespace: str
+    name: str
+    id: str
+    typename: str
 
 
-class Expiration(object):
-    class Mode(Enum):
-        AFTER_INVOKE = 0
-        AFTER_WRITE = 1
+class TypeSerializer(metaclass=abc.ABCMeta):
+    __slots__ = ()
 
-    def __init__(self, expire_after: timedelta, expire_mode: Mode=Mode.AFTER_INVOKE):
-        self.expire_mode = expire_mode
-        self.expire_after_millis = self.total_milliseconds(expire_after)
-        if self.expire_after_millis <= 0:
-            raise ValueError("expire_after_millis must be a positive number.")
+    """
+    A base class for a TypeSerializer. A TypeSerializer is responsible of serialising and deserializing a specific
+    value type.
+    """
 
-    @staticmethod
-    def total_milliseconds(expire_after: timedelta):
-        return int(expire_after.total_seconds() * 1000.0)
+    @abc.abstractmethod
+    def serialize(self, value) -> bytes:
+        """
+        Serialize the given value to bytes.
+
+        :param value: the value to serialize.
+        :return: a byte representation of the given value.
+        """
+        pass
+
+    @abc.abstractmethod
+    def deserialize(self, byte: bytes):
+        """
+        Deserialize a value from the given byte representation.
+
+        :param byte: the bytes that represent the value to deseralize.
+        :return: the deserialized value.
+        """
+        pass
 
 
-class AfterInvoke(Expiration):
-    def __init__(self, expire_after: timedelta):
-        super().__init__(expire_after, expire_mode=Expiration.Mode.AFTER_INVOKE)
+class Type(metaclass=abc.ABCMeta):
+    __slots__ = ("typename",)
+
+    """
+    A base representation of a Stateful Function value type.
+    """
+
+    def __init__(self, typename: str):
+        """
+        :param typename: a TypeName represented as a string of the form <namespace>/<name> of this type.
+        """
+        if not typename:
+            raise ValueError("typename can not be missing")
+        self.typename = typename
+
+    @abc.abstractmethod
+    def serializer(self) -> TypeSerializer:
+        """
+        :return: a serializer for this type.
+        """
+        pass
 
 
-class AfterWrite(Expiration):
-    def __init__(self, expire_after: timedelta):
-        super().__init__(expire_after, expire_mode=Expiration.Mode.AFTER_WRITE)
+class ValueSpec(object):
+    __slots__ = ("name", "type", "duration", "after_call", "after_write")
 
+    def __init__(self,
+                 name: str,
+                 type: Type,
+                 expire_after_call: timedelta = None,
+                 expire_after_write: timedelta = None):
+        """
+        ValueSpec a specification of a persisted value.
 
-class StateSpec(object):
-    def __init__(self, name, expire_after: Expiration=None):
-        self.name = name
-        self.expiration = expire_after
+        :param name: a user defined state name, used to represent this value. A name must be lower case, alphanumeric string
+        without spaces, that starts with either a letter or an underscore (_)
+        :param type: the value's Type. (see
+        statefun.Type) :param expire_after_call: expire (remove) this value if a call to this function hasn't been
+        made for the given duration.
+        :param expire_after_write: expire this value if it wasn't written to for the given duration.
+        """
         if not name:
-            raise ValueError("state name must be provided")
-
-
-class StateRegistrationError(Exception):
-    pass
-
-
-class StatefulFunction(object):
-    def __init__(self, fun, state_specs: List[StateSpec], expected_messages=None):
-        self.known_messages = expected_messages[:] if expected_messages else None
-        self.func = fun
-        if not fun:
-            raise ValueError("function code is missing.")
-        self.registered_state_specs = {}
-        if state_specs:
-            for state_spec in state_specs:
-                if state_spec.name in self.registered_state_specs:
-                    raise StateRegistrationError("duplicate registered state name: " + state_spec.name)
-                self.registered_state_specs[state_spec.name] = state_spec
-
-    def unpack_any(self, any: Any):
-        if self.known_messages is None:
-            return None
-        for cls in self.known_messages:
-            if any.Is(cls.DESCRIPTOR):
-                instance = cls()
-                any.Unpack(instance)
-                return instance
-
-        raise ValueError("Unknown message type " + any.type_url)
+            raise ValueError("name can not be missing.")
+        if not name.isidentifier():
+            raise ValueError(
+                f"invalid name {name}. A spec name can only contains alphanumeric letters (a-z) and (0-9), "
+                f"or underscores ( "
+                f"_). A valid identifier cannot start with a number, or contain any spaces.")
+        if iskeyword(name):
+            forbidden = '\n'.join(kwlist)
+            raise ValueError(
+                f"invalid spec name {name} (Python SDK specifically). since {name} will result as an attribute on "
+                f"context.store.\n"
+                f"The following names are forbidden:\n {forbidden}")
+        if not name.islower():
+            raise ValueError(f"Only lower case names are allowed, {name} is given.")
+        self.name = name
+        if not type:
+            raise ValueError("type can not be missing.")
+        if not isinstance(type, Type):
+            raise TypeError("type is not a StateFun type.")
+        self.type = type
+        if expire_after_call and expire_after_write:
+            # both can not be set.
+            raise ValueError("Either expire_after_call or expire_after_write can be set, but not both.")
+        if expire_after_call:
+            self.duration = int(expire_after_call.total_seconds() * 1000.0)
+            self.after_call = True
+            self.after_write = False
+        elif expire_after_write:
+            self.duration = int(expire_after_write.total_seconds() * 1000.0)
+            self.after_call = False
+            self.after_write = True
+        else:
+            self.duration = 0
+            self.after_call = False
+            self.after_write = False
 
 
 def parse_typename(typename):
-    """parses a string of type namespace/type into a tuple of (namespace, type)"""
+    """
+    Parse a TypeName string into a namespace, type pair.
+    :param typename: a string of the form <namespace>/<type>
+    :return: a tuple of a namespace type.
+    """
     if typename is None:
         raise ValueError("function type must be provided")
     idx = typename.rfind("/")
@@ -183,107 +157,54 @@ def parse_typename(typename):
     return namespace, type
 
 
-def deduce_protobuf_types(fn):
+def simple_type(typename=None, serialize_fn=None, deserialize_fn=None) -> Type:
     """
-    Try to extract the class names that are attached as the typing annotation.
+    Create a user defined Type, simply by providing two functions. One for serializing a value to bytes, and one for
+    deserializing the value from bytes.
+    For example:
 
-    :param fn: the function with the annotated parameters.
-    :return: a list of classes or None.
+    import json
+    tpe = simple_type("org.foo.bar/User", serialize_fn=json.dumps, deserialize_fn=json.loads)
+
+    this defines a StateFun type (Type) of the typename org.foo.bar/User and it is a JSON object.
+
+    :param typename: this type's TypeName.
+    :param serialize_fn: a function that is able to serialize values of this type.
+    :param deserialize_fn: a function that is able to deserialize bytes into values of this type.
+    :return: a Type definition.
     """
-    spec = inspect.getfullargspec(fn)
-    if not spec:
-        return None
-    if len(spec.args) != 2:
-        raise TypeError("A stateful function must have two arguments: a context and a message. but got ", spec.args)
-    message_arg_name = spec.args[1]  # has to be the second element
-    if message_arg_name not in spec.annotations:
-        return None
-    message_annotation = spec.annotations[message_arg_name]
-    if inspect.isclass(message_annotation):
-        return [message_annotation]
-    try:
-        # it is not a class, then it is only allowed to be
-        # typing.SpecialForm('Union')
-        return list(message_annotation.__args__)
-    except Exception:
-        return None
+    return SimpleType(typename, serialize_fn, deserialize_fn)
 
 
-class StatefulFunctions:
-    def __init__(self):
-        self.functions = {}
-
-    def register(self, typename: str, fun, state_specs: List[StateSpec]=None):
-        """registers a StatefulFunction function instance, under the given namespace with the given function type. """
-        if fun is None:
-            raise ValueError("function instance must be provided")
-        namespace, type = parse_typename(typename)
-        expected_messages = deduce_protobuf_types(fun)
-        self.functions[(namespace, type)] = StatefulFunction(fun, state_specs, expected_messages)
-
-    def bind(self, typename, states: List[StateSpec]=None):
-        """wraps a StatefulFunction instance with a given namespace and type.
-           for example:
-            s = StateFun()
-
-            @s.define("com.foo.bar/greeter")
-            def greeter(context, message):
-                print("Hi there")
-
-            This would add an invokable stateful function that can accept messages
-            sent to "com.foo.bar/greeter".
-         """
-
-        def wrapper(function):
-            self.register(typename, function, states)
-            return function
-
-        return wrapper
-
-    def for_type(self, namespace, type):
-        return self.functions[(namespace, type)]
+# ----------------------------------------------------------------------------------------------------------
+# Internal
+# ----------------------------------------------------------------------------------------------------------
 
 
-def kafka_egress_record(topic: str, value, key: str = None):
-    """
-    Build a ProtobufMessage that can be emitted to a Kafka generic egress.
+class SimpleType(Type):
+    __slots__ = ("typename", "_ser")
 
-    :param topic: The Kafka destination topic for that record
-    :param key: the utf8 encoded string key to produce (can be empty)
-    :param value: the Protobuf value to produce
-    :return: A Protobuf message representing the record to be produced via the Kafka generic egress.
-    """
-    if not topic:
-        raise ValueError("A destination Kafka topic is missing")
-    if not value:
-        raise ValueError("Missing value")
-    record = KafkaProducerRecord()
-    record.topic = topic
-    record.value_bytes = value.SerializeToString()
-    if key is not None:
-        record.key = key
-    return record
+    def __init__(self, typename, serialize_fn, deserialize_fn):
+        super().__init__(typename)
+        if not serialize_fn:
+            raise ValueError("serialize_fn is missing")
+        if not deserialize_fn:
+            raise ValueError("deserialize_fn is missing")
+        self._ser = UserTypeSerializer(serialize_fn, deserialize_fn)
 
-def kinesis_egress_record(stream: str, value, partition_key: str, explicit_hash_key: str = None):
-    """
-    Build a ProtobufMessage that can be emitted to a Kinesis generic egress.
+    def serializer(self) -> TypeSerializer:
+        return self._ser
 
-    :param stream: The AWS Kinesis destination stream for that record
-    :param partition_key: the utf8 encoded string partition key to use
-    :param value: the Protobuf value to produce
-    :param explicit_hash_key: a utf8 encoded string explicit hash key to use (can be empty)
-    :return: A Protobuf message representing the record to be produced to AWS Kinesis via the Kinesis generic egress.
-    """
-    if not stream:
-        raise ValueError("Missing destination Kinesis stream")
-    if not value:
-        raise ValueError("Missing value")
-    if not partition_key:
-        raise ValueError("Missing partition key")
-    record = KinesisEgressRecord()
-    record.stream = stream
-    record.value_bytes = value.SerializeToString()
-    record.partition_key = partition_key
-    if explicit_hash_key is not None:
-        record.explicit_hash_key = explicit_hash_key
-    return record
+
+class UserTypeSerializer(TypeSerializer):
+    __slots__ = ("serialize_fn", "deserialize_fn")
+
+    def __init__(self, serialize_fn, deserialize_fn):
+        self.serialize_fn = serialize_fn
+        self.deserialize_fn = deserialize_fn
+
+    def serialize(self, value):
+        return self.serialize_fn(value)
+
+    def deserialize(self, byte_string):
+        return self.deserialize_fn(byte_string)
